requirements.txt
loguru==0.7.2
selenium==4.8.0
yt-dlp==2024.7.16
yt-dlp-youtube-oauth2 @ https://github.com/coletdjnz/yt-dlp-youtube-oauth2/archive/refs/heads/master.zip
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1